requirements.txt
aiohttp
fastapi
pydantic
pydantic-settings
pyyaml
uvicorn